requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

# lev-bd/bot.py: 1,2,3,4,5,6
python_telegram_bot == 13.10
